封装

+++ Student.Core.API/Migrations/myDbContextModelSnapshot.cs
@@ -111,9 +111,9 @@ namespace Student.Core.API.Migrations
 
             modelBuilder.Entity("Student.Model.StudentInfo", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint")
+                        .HasColumnType("int")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
                     b.Property<string>("Address")
@@ -160,7 +160,7 @@ namespace Student.Core.API.Migrations
                     b.HasData(
                         new
                         {
-                            Id = 1L,
+                            Id = 1,
                             Address = "朝阳区朝阳公园西路9号院九号",
                             DepartId = 4,
                             Email = "xiaoan@stu.com",
@@ -175,7 +175,7 @@ namespace Student.Core.API.Migrations
                         },
                         new
                         {
-                            Id = 2L,
+                            Id = 2,
                             Address = "北京市朝阳区东三环中路甲10号",
                             DepartId = 3,
                             Email = "laoli@stu.com",
